No, let's get back to the color asthetics of the COBOL html coverage report. It's not terrible, but has almost an arcade/tron feel to it, it's certainly glowing like the 1980's and the tommyknockers movie.
let's make it more weighty like the font and style of this, image so someone that has worked in COBOL long ago is transported back in time, against their will. Perhaps there's a COBOL language syntax highlighter we could utilize for getting the basic language keywords highlighted, and then the font style matched.

And, of course, we still need to keep the indicators to verify the covered, not covered, partially covered, and non-executable.